[host] format string type fixes, %Ix doesn't exist in mingw

+++ host/Util.h
@@ -59,7 +59,7 @@ public:
       if (buffer[i] == '\n' || buffer[i] == '\r')
         buffer[i] = 0;
 
-    fprintf(stderr, "[E] %20s:%-4u | %-30s | %s: 0x%08x (%s)\n", file, line, function, desc, status, buffer);
+    fprintf(stderr, "[E] %20s:%-4u | %-30s | %s: 0x%08x (%s)\n", file, line, function, desc, (int)status, buffer);
     LocalFree(buffer);
   }
